Function ic_crypto_internal_fs_ni_dkg::forward_secure::baby_giant

pub fn baby_giant(
    tgt: &FP12, 
    base: &FP12, 
    lo: isize, 
    range: isize
) -> Option<isize>

Solves discrete log problem with baby-step giant-step.

Returns: find (\x -> base^x == tgt) [lo..lo + range - 1]

using an O(sqrt(N)) approach rather than a naive O(N) search.

We call reduce() before every tobytes() because this algorithm requires the same element to serialize identically every time. (MIRACL does not automatically perform Montgomery reduction for serialization, so in general x == y does not imply x.tobytes() == y.tobytes().)

We cut the exponent in half, that is, for a range of 2^46, we build a table of size 2^23 then perform up to 2^23 FP12 multiplications and lookups. Depending on the cost of CPU versus RAM, it may be better to split differently.