Function ic_crypto_internal_fs_ni_dkg::forward_secure::verify_ciphertext_integrity

pub fn verify_ciphertext_integrity(
    crsz: &CRSZ, 
    tau: &[Bit], 
    associated_data: &[u8], 
    sys: &SysParam
) -> Result<(), ()>

Verify ciphertext integrity

Part of DVfy of Section 7.1 of https://eprint.iacr.org/2021/339.pdf In addition to verifying the proofs of chunking and sharing, we must also verify ciphertext integrity.