Module ic_crypto_internal_fs_ni_dkg::nizk_chunking

Proofs of correct chunking

Structs

ChunkingInstance
Instance for a chunking relation.

ChunkingWitness
Witness for the validity of a chunking instance.

ProofChunking
Zero-knowledge proof of chunking.

Enums

ZkProofChunkingError
Creating or verifying a proof of correct chunking failed.

Constants

CHALLENGE_BITS
Defined as ceil(SECURITY_LEVEL/NUM_ZK_REPETITIONS)

NUM_ZK_REPETITIONS
The number of parallel proofs handled by one challenge

Functions

prove_chunking
Create a proof of correct chunking

verify_chunking
Verify a proof of correct chunking