Module ic_crypto_internal_fs_ni_dkg::random_oracles

Hashing to group elements (fields, curves)

Structs

HashedMap
Ordered map storing the unique digests of values using unique digests as the keys.

Traits

UniqueHash
Hashes the unique encoding of some structured data. Each data type uses a distinct domain separator.

Functions

random_oracle
Computes the hash of a struct using an hash function that can be modelled as a random oracle.

random_oracle_to_miracl_g1
Computes the hash of a struct using an hash function that can be modelled as a random oracle. Returns a group element of G1 in BLS12_381.

random_oracle_to_scalar
Computes the hash of a struct using an hash function that can be modelled as a random oracle. Returns an element in the scalar field of curve BLS12_381.

Type Definitions

HashableMap
Ordered map with elements that can be uniquely hashed.