Module ic_crypto_internal_fs_ni_dkg::utils

Utility functions for the NI-DKG code

Structs

RAND_ChaCha20
A random number generator based on the ChaCha20 stream cipher

Functions

big_one
Identity element in the scalar field of curve BLS12_381.

big_zero
Zero element in the scalar field of curve BLS12_381.

curve_order
Order of the prime order subgroup of curve BLS12_381.

ecp2_inf
Point at infinity on G2 of curve BLS12_381.

ecp_inf
Point at infinity on G1 of curve BLS12_381.

field_add
Addition of two field elements modulo the prime order of the group.

field_mul
Multiplication of two field elements modulo the prime order of the group.

htp2_bls12381
Hash a message onto the BLS12-381 G2 curve

oracle_p
Hash a message to a random integer modulo the BLS12-381 order

process_ecp
Feeds the standard serialisation of an element of G1==ECP to a hash.

process_ecp2
Feeds the standard serialisation of an element of G2==ECP2 to a hash.

process_fr
Feeds an element of Z_p BIG to a hash.